Remove simple search from navigation.
The navigation is not shown by default. However, compared to the classic
theme, this theme has moved the simple search from the navigation into
the sidebar. If the navigation is included/enabled, we should not have a
duplicate simple search in the navigation and sidebar. Hence its removal
from the navigation.